Changed namespace structure
