+++ KnueppelKampfBase/Networking/Packets/ClientPackets/KeepAlivePacket.cs
@@ -2,16 +2,16 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace KnueppelKampfBase.Networking.Packets
+namespace KnueppelKampfBase.Networking.Packets.ClientPackets
 {
-    class DeclineConnectPacket : Packet
+    class KeepAlivePacket : SaltedPacket
     {
-        public DeclineConnectPacket()
+        public KeepAlivePacket(byte salt) : base(salt)
         {
 
         }
 
-        public DeclineConnectPacket(byte[] bytes) : base(bytes)
+        public KeepAlivePacket(byte[] bytes) : base(bytes)
         {
 
         }
