+++ KnueppelKampfBase/Networking/Packets/ServerPackets/FullWorldPacket.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace KnueppelKampfBase.Networking.Packets
+namespace KnueppelKampfBase.Networking.Packets.ServerPackets
 {
     /// <summary>
     /// Packet containing the entire world state, sent by server
